prism-web/prism-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
MarkupSafe==3.0.2
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
passlib==1.7.4
portalocker==2.10.1
prettytable==3.16.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyecharts==2.0.8
pymongo==4.13.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
xxhash==3.5.0
zstandard==0.23.0
